Add application status and interview type chart to metrics
