+++ app/profile/page.tsx
@@ -4,16 +4,12 @@ import React, { useEffect, useState } from "react";
 import { useSession } from "next-auth/react";
 import useSWR, { mutate } from "swr";
 import axios from "axios";
-import { Chart } from "chart.js/auto";
 import SuggestedSkillsCard from "../components/profile/SuggestedSkillsCard";
 import ProfileCard from "../components/profile/ProfileCard";
 import UserSkillsCard from "../components/profile/UserSkillsCard";
 import JobRejectionCard from "../components/profile/JobRejectionCard";
 import JobOfferCard from "../components/profile/JobOfferCard";
 import getUserJobPostings from "../lib/getUserJobPostings";
-import getUserJobInterviews from "../lib/getUserJobInterviews";
-import { getJobsByApplicationStatus } from "../lib/getJobsByApplicationStatus";
-import { convertToSentenceCase } from "../lib/convertToSentenceCase";
 import { RejectionInitiator } from "@prisma/client";
 import UpcomingInterviews from "../components/profile/UpcomingInterviews";
 
@@ -76,18 +72,6 @@ interface Rejection {
   };
 }
 
-const interviewTypes = [
-  { type: "FOLLOW_UP", label: "Follow Up", color: "#94a3b8" },
-  { type: "PHONE_SCREEN", label: "Phone Screen", color: "#facc15" },
-  { type: "ON_SITE", label: "On Site", color: "#4ade80" },
-  { type: "VIDEO_INTERVIEW", label: "Video Interview", color: "#818cf8" },
-  { type: "INTERVIEW", label: "Interview", color: "#9ca3af" },
-  { type: "ASSESSMENT", label: "Assessment", color: "#f472b6" },
-  { type: "TECHNICAL", label: "Technical", color: "#c084fc" },
-  { type: "PANEL", label: "Panel", color: "#60a5fa" },
-  { type: "FINAL_ROUND", label: "Final Round", color: "#f87171" },
-];
-
 const fetcher = async (url: string, options: RequestInit) => {
   const response = await fetch(url, options);
   return response.json();
@@ -119,12 +103,6 @@ function Profile() {
   const jobInterviews = userInterviews || [];
 
   const [jobPostings, setJobPostings] = useState<JobPosting[]>([]);
-  const [statusPercentages, setStatusPercentages] = useState<
-    Map<string, number>
-  >(new Map());
-  const [interviewTypeFrequency, setInterviewTypeFrequency] = useState<
-    Record<string, number>
-  >({});
 
   useEffect(() => {
     async function fetchJobPostings() {
@@ -148,205 +126,6 @@ function Profile() {
         : []
     )
   );
-
-  useEffect(() => {
-    async function fetchData() {
-      try {
-        const { percentages } = await getJobsByApplicationStatus();
-        setStatusPercentages(percentages);
-      } catch (error) {
-        console.error("Error fetching application status:", error);
-      }
-    }
-    fetchData();
-  }, []);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const { interviewTypeFrequency } = await getUserJobInterviews();
-        setInterviewTypeFrequency(interviewTypeFrequency);
-      } catch (error) {
-        console.error("Error fetching user job interviews:", error);
-      }
-    };
-    fetchData();
-  }, []);
-
-  useEffect(() => {
-    // Draw application status chart
-    const chartContainer = document.getElementById("applicationStatusChart");
-    if (chartContainer instanceof HTMLCanvasElement) {
-      const applicationStatuses = [
-        "Saved",
-        "Applied",
-        "Interview",
-        "Offer",
-        "Rejected",
-      ];
-      const labels = applicationStatuses.map(
-        (applicationStatus) => applicationStatus
-      );
-      const datasets = [
-        {
-          label: "Job Percentage",
-          data: labels.map((label) => statusPercentages.get(label) || 0),
-          backgroundColor: [
-            "rgba(255, 99, 132, 0.2)",
-            "rgba(54, 162, 235, 0.2)",
-            "rgba(255, 206, 86, 0.2)",
-            "rgba(75, 192, 192, 0.2)",
-            "rgba(153, 102, 255, 0.2)",
-          ],
-          borderColor: [
-            "rgba(255, 99, 132, 1)",
-            "rgba(54, 162, 235, 1)",
-            "rgba(255, 206, 86, 1)",
-            "rgba(75, 192, 192, 1)",
-            "rgba(153, 102, 255, 1)",
-          ],
-          borderWidth: 1,
-        },
-      ];
-
-      const chart = new Chart(chartContainer, {
-        type: "bar",
-        data: {
-          labels: labels,
-          datasets: datasets,
-        },
-        options: {
-          plugins: {
-            legend: { display: false },
-            tooltip: {
-              callbacks: {
-                label: function (context) {
-                  let label = context.dataset.label || "";
-                  if (label) label += ": ";
-                  if (context.parsed.y !== null) {
-                    const value = context.parsed.y;
-                    label += Number.isInteger(value)
-                      ? value.toFixed(0) + "%"
-                      : value.toFixed(2) + "%";
-                  }
-                  return label;
-                },
-              },
-            },
-          },
-          responsive: true,
-          maintainAspectRatio: false,
-          scales: {
-            y: {
-              beginAtZero: true,
-              title: { display: true, text: "Percentage", color: "white" },
-              ticks: { color: "white", maxTicksLimit: 5 },
-            },
-            x: {
-              title: {
-                display: true,
-                text: "Application Status",
-                color: "white",
-              },
-              ticks: { color: "white" },
-            },
-          },
-          layout: { padding: { left: 10, right: 10, top: 10, bottom: 10 } },
-        },
-      });
-
-      return () => {
-        chart.destroy();
-      };
-    }
-  }, [statusPercentages]);
-
-  useEffect(() => {
-    if (Object.keys(interviewTypeFrequency).length > 0) {
-      // Draw chart when interviewTypeFrequency changes
-      const chartContainer = document.getElementById(
-        "interviewTypeFrequencyChart"
-      );
-      if (chartContainer) {
-        const chartColors: { [key: string]: string } = {};
-        interviewTypes.forEach((interview) => {
-          chartColors[interview.type] = interview.color;
-        });
-
-        const labels = interviewTypes.map((interview) =>
-          convertToSentenceCase(interview.label)
-        );
-        const backgroundColor = interviewTypes.map(
-          (interview) => chartColors[interview.type]
-        );
-        const data = interviewTypes.map(
-          (interview) => interviewTypeFrequency[interview.type] || 0
-        );
-
-        const chart = new Chart(chartContainer as HTMLCanvasElement, {
-          type: "bar",
-          data: {
-            labels: labels,
-            datasets: [
-              {
-                label: "Interview Type Frequency",
-                data: data,
-                backgroundColor: backgroundColor,
-              },
-            ],
-          },
-          options: {
-            responsive: true,
-            maintainAspectRatio: false,
-            scales: {
-              x: {
-                title: {
-                  display: true,
-                  text: "Interview Type",
-                  color: "#fff",
-                },
-                ticks: {
-                  color: "#fff",
-                  font: {
-                    size: 10,
-                  },
-                },
-              },
-              y: {
-                title: {
-                  display: true,
-                  text: "Frequency",
-                  color: "#fff",
-                },
-                ticks: {
-                  color: "#fff",
-                },
-                beginAtZero: true,
-              },
-            },
-            plugins: {
-              tooltip: {
-                enabled: true,
-                callbacks: {
-                  label: (tooltipItem) => {
-                    const value = tooltipItem.formattedValue;
-                    return `Frequency: ${value}`;
-                  },
-                },
-              },
-              legend: {
-                display: false,
-              },
-            },
-          },
-        });
-
-        return () => {
-          chart.destroy();
-        };
-      }
-    }
-  }, [interviewTypeFrequency]);
 
   const handleDeleteRejection = async (id: string) => {
     try {
@@ -381,17 +160,6 @@ function Profile() {
       <div className="mt-5">
         <UpcomingInterviews jobInterviews={jobInterviews} />
       </div>
-      <div className="">
-        <div className="w-full h-[550px] mt-2">
-          <canvas id="applicationStatusChart"></canvas>
-        </div>
-      </div>
-
-      {Object.keys(interviewTypeFrequency).length > 0 && (
-        <div className="w-full h-[550px] mt-2">
-          <canvas id="interviewTypeFrequencyChart"></canvas>
-        </div>
-      )}
 
       <div className="mt-4 grid grid-cols-1 gap-3 sm:grid-cols-2 lg:grid-cols-3">
         {jobRejections.map((rejection: Rejection) => (
